requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
yookassa>=3.0.0
python-dotenv>=1.0
reportlab>=3.6.12
matplotlib>=3.8.0
python-telegram-bot[job-queue]>=20.0

